+++ packages/core/src/features/gen3/index.ts
@@ -1,17 +1,8 @@
 import {
   gen3Api,
-  useGetCSRFQuery,
-  selectCSRFToken,
-  selectCSRFTokenData,
-  selectHeadersWithCSRFToken,
-  type CSRFToken,
+
 } from './gen3Api';
 
 export {
-  gen3Api,
-  useGetCSRFQuery,
-  selectCSRFToken,
-  selectCSRFTokenData,
-  selectHeadersWithCSRFToken,
-  type CSRFToken,
+  gen3Api
 };
