+++ packages/sampleCommons/.env.development
@@ -1,11 +1,11 @@
-GEN3_COMMONS_NAME=bih
-NEXT_PUBLIC_GEN3_API='https://localhost'
+GEN3_COMMONS_NAME=gen3
+NEXT_PUBLIC_GEN3_API='https://localhost:3010'
 #NEXT_PUBLIC_GEN3_SUBMISSION_API='https://localhost/api/v0/submission'
 #NEXT_PUBLIC_GEN3_FENCE_API='https://locahost
 #NEXT_PUBLIC_GEN3_DOMAIN=https://localhost
 #NEXT_PUBLIC_GEN3_MDS_API=https://localhost/mds
 #NEXT_PUBLIC_GEN3_AI_SEARCH_API=https://localhost/ai
-NEXT_PUBLIC_GEN3_GUPPY_API=https://imaging-hub.data-commons.org/guppy
+#NEXT_PUBLIC_GEN3_GUPPY_API=https://imaging-hub.data-commons.org/guppy
 #NEXT_PUBLIC_GEN3_FENCE_API=https://localhost
 #NEXT_PUBLIC_GEN3_AUTHZ_API=https://localhost/authz
 #NEXT_PUBLIC_GEN3_REDIRECT_URL=''
